+++ changes.md
@@ -1,3 +1,10 @@
 New version is based on Podman
 
 Drop Ollama support
+
+Updated:
+- llama.cpp
+- KoboldCPP
+- Text generation web UI
+- ComfyUI 
+- ComfyUI-Manager